requirements.txt
python-binance==1.0.19
python-telegram-bot==20.7
aiohttp==3.9.1
websockets==12.0
python-dotenv==1.0.0
asyncio==3.4.3
pydantic==2.5.2
loguru==0.7.2
aiosmtplib==3.0.1
email-validator==2.1.0